feat(build): allow build in release with debug symbols
This set of patch make possible to build most bin packages in release with debugging symbols allowing us to have fast binaries but still able to have a backtrace in case of errors.

Most of the bin packages will be built with debug symbols, but some of them do not support this configuration yet. In this case, they will be built in release mode, without debugging symbols.