feat(numpy): remove numpy
This outdated Numpy do not build since a lot of time (before python 3 adoption) and is not really maintained for Fw4spl applications. We need to rethink the way we add it to our python installation and reenable it when it is ready. For now, as it breaks build, we choose to remove it